+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Anticipated release: June 1, 2020
 #### 🚀 New features
 
 - Update Print Preview Layout ([#2182])
+- Make the "Saving" message persist for 750ms ([#2207])
 
 #### 🐛 Bugs fixed
 
@@ -21,4 +22,5 @@ See our [release history](https://github.com/18F/cms-hitech-apd/releases)
 [#2176]: https://github.com/18F/cms-hitech-apd/issues/2176
 [#2179]: https://github.com/18F/cms-hitech-apd/issues/2179
 [#2182]: https://github.com/18F/cms-hitech-apd/issues/2182
+[#2207]: https://github.com/18F/cms-hitech-apd/issues/2207
 [#2206]: https://github.com/18F/cms-hitech-apd/issues/2206